requirements.txt
# BreastEdge AI - Python Dependencies

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0

# Medical AI
monai>=1.3.0

# Image processing
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0

# Web interface
gradio>=6.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional (for development)
jupyter>=1.0.0
ipython>=8.12.0
